MB-55904 Enable spilling for window functions
This removes the restriction that ORDER BY operations that don't permit
clipping values aren't eligible for spilling.

GetFreeMemPercent aims to have a finer grain estimate of current system
state so we can better react to sudden spikes.

When we're not clipping, the spilling is compressed using zlib; this
trades temp file space for CPU usage/response time.

Change-Id: I553e3c6edf46a253d2fae4e714d4e671cacf2da0
Reviewed-on: https://review.couchbase.org/c/query/+/190568
Reviewed-by: Marco Greco <marco.greco@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>