MB-56787 Proper handling of scalar value used as array in index creation
We currently have an inconsistency for array index keys between two
equivalent ways of defining the array index key:
 - simplified form - (ALL arr1)
 - long form       - (ALL ARRAY v FOR v IN arr1 END)
the difference shows up when the arr1 field in a document contains
a scalar value (i.e. not an actual array). For example:
  {"arr1": 1}
in this case for the long form, we index a NULL value for the array
index key, which is correct; but for the simplified form, we
implicitly convert the scale value into an array of 1 element, and
index [1]. This is incorrect in that if we use a covering index
scan, for either UNNEST or an ANY clause:
  FROM default d UNNEST d.arr1 AS v WHERE v > 0
  FROM default d WHERE ANY v IN d.arr1 SATISFIES v > 0 END
in both cases if covering index scan is used, we end up returning
the document with the scalar value, vs if primary scan or sequential
scan is used, we do not return that document.

The difference is in how we treat non-array values in the
EvaluateForIndex() function of All expression (expression.All). In
the long form, we have an Array expression (expression.Array) and
the EvaluateForIndex() function for Array returns null for a
non-array value, and subsequently the All expression returns
_NULL_ARRAY; in the simplified form, we evaluate the expression
directly and for non-array values we return an array with the
scalar value.

This behavior has been there since the simplified form of array index
key specification was first introduced, and it affects both ALL and
DISTINCT versions of the array index key specification. There was an
older bug on this: MB-38997 that was never fixed. The inconsistent
behavior, although it should hopefully be rare (only occurs when
scalar value used for an array field), should be fixed, and for
existing customer indexes, there is no easy way to address this
except for asking the customer to drop and recreate the index if
this issue surfaces.

The fix is to check whether simplified form of the array index key
is used, and if so return _NULL_ARRAY instead of an array of 1
element.

Change-Id: Ib7bf82765980d860fb0346d376eeb7f2dc57b6e5
Reviewed-on: https://review.couchbase.org/c/query/+/190669
Reviewed-by: Donald Haggart <donald.haggart@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>