MB-56524: Support for EXPLAIN FUNCTION statement.
This statement will allow users to see query plans of N1QL queries inside UDFs.

Syntax :   EXPLAIN FUNCTION func_name;
Example: EXPLAIN FUNCTION func1;

Created a new AST node, execution operator called ExplainFunction.

The privileges required to run EXPLAIN FUNCTION are the necessary execute function permissions ( either global or scope )

The query plans of N1QL statements inside UDFs are built in the Execution phase. And appended to an array of query plans. And are then marshalled.

For Inline Function:

If the body of the function has a subquery in it.
We build the plan of the query.

For JS Functions:

For now, if EXPLAIN FUNCTION is executed on a JS Function - a E_FUNCTIONS_UNSUPPORTED_ACTION error is thrown , as we depend on js-evaluator for a method to return a list of statements inside a JS UDF.

But this PR adds support for it.

For each statement in this list:
1. Build the query plan
2. Authorisation check for privileges for the statement

Note - Explaining is disabled on EXPLAIN, EXECUTE, ADVISE statements - and instead we display a “Explain is not supported” text.
	  For PREPARE statements, we build the query plan of the actual statement being prepared instead.

Change-Id: I28dd611b5805f8c173b0bc4bf20a846f4ffde0ef
Reviewed-on: https://review.couchbase.org/c/query/+/190749
Reviewed-by: Marco Greco <marco.greco@couchbase.com>
Tested-by: Dhanya Gowrish <dhanya.gowrish@couchbase.com>