MB-56727 Proper determination of covering for unnest scan
As part of changes for MB-32506, we modified the mechanism
to cover unnest index scans with array index keys. For
the unnest expression(s), we used to skip these expressions
in covering check, now we include these expression(s) in
covering expressions, which then allowed the same expression
appearing in other parts of the query (e.g. projection clause)
to be covered as well, which is incorrect.

The fix is to return to previous mechanism, and skip covering
checks for unnest expression(s) only, since we are doing this
on the actual expression pointer, this only affect the actual
unnest expression(s), and not the same expression in other
parts of the query.

As part of this change, the newly added code (from MB-32506)
for unnesting on keyspace only is no longer needed (included
in the check) and thus removed.

Change-Id: I3bee6cb20aa1da97542b7c3c4fa40bfa16319799
Reviewed-on: https://review.couchbase.org/c/query/+/190645
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-on: https://review.couchbase.org/c/query/+/190655
Well-Formed: Restriction Checker
Reviewed-on: https://review.couchbase.org/c/query/+/191045