MB-56940 Detect and mark "constant" filters in subterms of OR-clause
The issue was first identified in 7.0.5 (MB-53793) and not in
later code lines since we handle OR-clause differently in later
codelines. However, we've identified issues with OR-clause
handling and fix the issues with MB-56039, after which the
symptom of MB-53793 now shows in 7.2.0 and 7.5.0.

The problematic query has an OR-clause like:

(c2 = 6 AND IFMISSINGORNULL($qp1, "") != "") OR (IFMISSINGORNULL($qp1, "") == "")

the issue is we have part of the predicate in each arm of the
OR-clause as "constant" predicate, i.e. it does not depend on
the document, rather it only depends on a named parameter.
Such subclauses causes issues when sarging index keys since
they don't generate any index spans but they are not detected
by the "covering" test we currently use to check whether the
index spans are "exact". The issue is somewhat similar to
MB-50071 except the "constant" predicate in this case is part
of subterms of an OR-clause and thus not detected by the fix
for MB-50071.

To fix this issue, we need to add a mechanism to detect such
"constant" predicates in subterms of an OR-clause. This is
built on top of the fix for MB-42474, and we check in case of
OR-clause whether any subterm of the OR has such "contant"
predicates, if so, mark an extra flag to prevent index
aggregation pushdown.

An extra modification is also done for classifying expressions,
such that when we have an OR clause under and AND, we do the
necessary OR-clause processing to remove any constant subterms.

Change-Id: Icef39a8a0ca1f1a25e6dd62da62197b798e68159
Reviewed-on: https://review.couchbase.org/c/query/+/189176
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-on: https://review.couchbase.org/c/query/+/189177
Well-Formed: Restriction Checker
Reviewed-on: https://review.couchbase.org/c/query/+/191066