MB-56956 Don't close operators more than once
Once close() has been called on an operator a concurrently executing
routine may be able to pool it.  Any further calls to close may corrupt
new use or be affected by re-initialisation (and typically panic) and as
such should be avoided.

Because of the varying nature of when operators may be used & reused,
preventing multiple calls to close() directly is often problematic.

So once an operator's inactive() function is called (on the first pass
through close()) it is flagged as such and further calls to close()
on it become no-ops.

This is imperfect - delayed second calls to close() could occur after
re-initialisation - at which point the inactive flag has been reset but
it does reduce the opportunity for such problems.

Where possible cases for multiple close() calls have been identified and
can be avoided, they are. It is often the case here that the operator's
output has been handed on to children and is not used directly by the
operator itself so once assigned is saved (in case of reopening) and set
to nil.

Change-Id: Ib7b2b5c42adac4497eea04cbe2c757be2642ecee
Reviewed-on: https://review.couchbase.org/c/query/+/191478
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Marco Greco <marco.greco@couchbase.com>
Tested-by: Donald Haggart <donald.haggart@couchbase.com>