MB-57024 Handle FALSE WHERE clause better
If the query has a FALSE WHERE clause, such as:
  WHERE 1 = 0
then we simplify the WHERE clause and set a flag indicating
that we have a FALSE WHERE clause, subsequently, we could
simplify query processing and generate an "empty" query plan
and avoid most processing at runtime. However, this currently
is only done for a KeyspaceTerm. In case we have a USE KEYS
clause, or if we have an ExpressionTerm/SubqueryTerm, we do
not currently check the flag and thus could produce wrong
results (since the original WHERE clause no longer exists).

The fix is to add checks for the FALSE WHERE clause flag in
several situations:
  - for USE KEYS clause, generate a KeyScan with an empty
    array instead of the originally specified keys
  - for ExpressionTerm/SubqueryTerm, generate an empty plan
    similar to how it is done for KeyspaceTerm

Change-Id: I501e9acf394dbfea393c0ffa3b884b72aa1141be
Reviewed-on: https://review.couchbase.org/c/query/+/191238
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-on: https://review.couchbase.org/c/query/+/191306
Well-Formed: Restriction Checker