MB-57038 Address memory usage for host to UUID mapping
This addresses the primary problem that we weren't closing off the copy
of the pool loaded to refresh the name to UUID mapping.  At the same
time, a full pool load is overkill here as we need solely three fields:
name, uuid & status.  So the fix aims to obtain these without the
unnecessary overheads of a full pool load.

At the same time, we had the opportunity for multiple requests to all be
trying to refresh at the same time needlessly wasting resources doing
so, so a lightweight gate is introduced to ensure only one refresh is
active at a time.

Lastly, since the query nodes are cached and they're used for the name
to uuid mapping, the map of host names and uuids has been reversed so
the uuid is the key and the name the payload in order to simplify the
more frequent use of the map.

Change-Id: I58a584751b6b0b047db110c78f18244dec0070c4
Reviewed-on: https://review.couchbase.org/c/query/+/191531
Reviewed-by: Marco Greco <marco.greco@couchbase.com>
Tested-by: Donald Haggart <donald.haggart@couchbase.com>