MB-57102 Use a separate structure for WITH clause
The WITH clause is currently represented by expression.Bindings,
which works ok for current WITH clause functionality, however
with planned enhancement of recursive WITH clause it becomes
cumbersome to keep using expression.Bindings to represent WITH
since now there are additional information that need to be kept.

The WITH clause is a fairly specific construct and making it a
generic expression type (which then requires all the handling of
a generic expression type) may not make sense in most situations.
Instead, we can create a new algebra structure specific for WITH
(algebra.With), and a corresponding interface in expression
(expression.With, together with expression.Withs). We can then
se that structure/type instead of expression.Bindings to
represent WITH clauses. This allows easy extension of the WITH
functionality.

Change-Id: Id976012e478f1a63ec2da36733713826802b7961
Reviewed-on: https://review.couchbase.org/c/query/+/191553
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Marco Greco <marco.greco@couchbase.com>