MB-57052 check bucket cleanup status every minute
I have had a look at both Sitaram's and Donald's suggestion, and having a look at buckets in the server stats loop might be potentially very expensive if we are expected to grow to up to 800 tenants per node,
and Donald's suggestion suffers from the same issues as the current implementation, namely being too hot on time.Now()

I have implemented a ticker that starts as soon as the memory manager starts a tenant, and checks every minute.
It has the same cost as the current implementation in terms of timers, maps and goroutines, and fewer time.Now() calls.


Change-Id: Icb51c29eed55f31aaf2f08440399362dbf5a8504
Reviewed-on: https://review.couchbase.org/c/query/+/191601
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Donald Haggart <donald.haggart@couchbase.com>