MB-57169 Retry host to node ID resolution...
... during topology manager initialisation.  This is to avoid the case
where we have "known Query nodes" without node IDs which can lead to
incorrect invocation of graceful shutdown.  We retry endlessly as this
is an independent go-routine and it is better to have no initial
topology awareness than be missing node IDs.  If a topology change
prepare is received before the initial node list is formed, the data
provided in the prepare operation will be used and the initialisation
will detect this and terminate.

To further protect against incorrectly invoking graceful shutdown, the
test if a node remains in the cluster has been extended to compare host
names if the existing definition lacks a node ID (and will then consider
matching host names sufficient).  A possible side-effect is to not
attempt graceful shutdown on a node, but it will still be ejected &
killed by ns_server.

Another addition is to process the prepare operation's EjectNodes list
and add any nodes listed therein that weren't already determined to be
being ejected (by comparing the known topology against the KeepNodes
list) to the manager's node ejection list.  This just to extend the
possibility of graceful shutdown to any of these nodes in the unlikely
event that any should ever exist.

Change-Id: Icd5779615bf5f38817753db28b15ca3b3ceb3465
Reviewed-on: https://review.couchbase.org/c/query/+/191820
Well-Formed: Restriction Checker
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>