MB-57279 Detect and mark CTE variable as correlated
The binding expression for a CTE could be correlated if the
CTE is inside a subquery. Current correlation is not detected
which could result in the subbquery itself not properly marked
as correlated.

Add logic to detect and set correlated variable identifiers,
and set a subquery to be correlated if such correlated
variable exists.

In addition, if the CTE variable is used on the right-hand
side of an IN-list, avoid using hash table optimization if
the CTE variable is correlated.

Change-Id: Ib61569896e541a315a34c8fe5f4c14c26cc4fedb
Reviewed-on: https://review.couchbase.org/c/query/+/175058
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>
Reviewed-on: https://review.couchbase.org/c/query/+/192193
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Well-Formed: Restriction Checker