MB-57251 Proper detect and mark correlated subqueries from encoded plan
With multiple query nodes, a prepared statment prepared from one
query node can be executed from another query node. This "transfer"
of a prepared statement is done from transfer and subsequent
interpretation of the encoded plan of the prepared statement.

This encoded plan currently contains the plan operator tree of the
main query plan, however, if a correlated subquery is used, it is
included as the query text with the keyword CORRELATED prefixed to
the subquery text. At execution time the subquery plan is built
and executed.

One potential issue with this is that for a correlated subquery,
when the plan is being built, it does not contain all the required
information, e.g., for each AST tree node, it does not contain
information on correlation. The correlation information is deduced
dudring formalization process, which is done after the original
statment is being parsed, but not done when a subquery is being
parsed since that is done as an expression, not a statement.

The CORRELATED keyword was originally added for a subquery
expression to address this issue. When parsing an expression, if
CORRELATED keyword is encounterd, it knows the subquery is
correlated (without the steps to actually detect the correlation),
and mark it as so.

The issue is that over the years, we've introduced other aspects
of handling correlation, e.g., we now mark an ExpressionTerm and
sometimes KeyspaceTerm (containing a USE KEYS clause) as
correlated. In case of ExpresionTerm, if it is not marked as
correlated, the runtime ExpressionScan will cache the result of
first execution and use that subsequently. Thus if it is not
properly marked, we'll not get the correct result, which is what's
seen in the repro.

The fix has several aspects:

 - in the parser, when CORRELATED keyword is encounterd, instead
   of just marking the subquery as correlated, it now goes through
   the subquery and go through the AST tree and mark all the nodes
   as "correlated", to be on the safe side.
 - a separate issue identified during investigation: a number of
   plan operators cannot be properly generated from encoded plan
   since these operators are not properly registered as valid
   operators. We need to properly register all plan operators
   such that they can be regenerated from encoded plan.

Change-Id: If70758cdb77a1dfa02e14eca92d3b86d2f19b052
Reviewed-on: https://review.couchbase.org/c/query/+/192054
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>
Reviewed-by: Marco Greco <marco.greco@couchbase.com>
Reviewed-by: Donald Haggart <donald.haggart@couchbase.com>
Reviewed-on: https://review.couchbase.org/c/query/+/192200
Well-Formed: Restriction Checker
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>