MB-57197 Fix unit test failure
The previous changes to this defect include a change that
"compensate" the date by timezone first before the truncation
operation, since time.Truncate() function always truncate on
UTC time. This change allows the truncation on local time,
e.g. DATE_TRUNC_STR("2023-06-10T12:15:04", "year") produces
"2023-01-01T00:00:00".

The issue with previous change is that this "compensation"
is done after various date manipulations, e.g. if truncate
by year, it first truncate by day, then substract the number
of days since 01-01 to get to the beginning of the year. If
the compensation is done after this manipulation, we may end
up with something that's a day off.

The fix is to move the logic for this "compensation" to the
place where the actual trunction occurs, but before other
poential manipulations.

The N1QL function DATE_TRUNC_MILLIS also uses the same
underlying truncation function, however in that case we
first convert milliseconds to a date, which should be in
UTC format before the truncation since the converted date
is (implicitly) in UTC thus the truncation should happen
in UTC as well.

Change-Id: I8c1e9c835dd0b43dddf036b856f232a47b61063b
Reviewed-on: https://review.couchbase.org/c/query/+/192433
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Marco Greco <marco.greco@couchbase.com>