MB-57346 Handle OR clause with IS MISSING better in index advisor
For queries with OR clause predicates, the index advisor first generate
individual structures for each arm of the (DNF-generated) OR clause,
then attempt to find common leading index keys among all arms. In this
process, if the same field is used in IS MISSING for one arm, and
something else (e.g. EQ or LT, etc) in another arm, then we can
"combine" the arms and use this field as a common leading index key.
The issue with this approach is that if this field is generated as the
leading index key, without a INCLUDE MISSING specification this index
cannot be used for the arm of the OR clause that has a IS MISSING
predicate on this field. (Note this logic was from the initial
implementation of the index advisor, before INCLUDE MISSING feature
was introduced).

The index advisor already keeps track of IS MISSING predicates
separately, however if we end up "combining" the predicate in the
process described above for OR clauses, we lost the information that
IS MISSING was used for a particular field (index key). To fix this,
we need to remember that an index key has IS MISSING predicate for
some arm of the OR clause, and in case this index key is generated
as the leading index key, we need to include INCLUDE MISSING in the
index key specification.

The fix includes the following:

 - add a new "wasMissing" field to predicate info structure, in
   addition to the existing "isMissing" field
 - handle the new "wasMissing" field when copying, etc
 - IS MISSING predicate is kept in "isMissing" initially, then
   during the merge process for OR subarms, if we end up removing
   a field from isMissing (combined with other predicate), add it
   in "wasMissing" to remember the presence of IS MISSING on this
   field (index key)
 - when generating common leading keys for OR clause, first try to
   use an index key that is not in "wasMissing" list; if this is
   not possible, then set and remember that we need to generate
   the index with leading key INCLUDE MISSING
 - this logic needs to be done for both rule-based and cost-based
   advise
 - previously there is special logic to handle the case where all
   the arms of OR has IS MISSING for a particular field; with
   INCLUDE MISSING we can relax this restriction a bit such that
   if there is no other predicates, allow such arm to be combined
   into common leading index key for OR, and generate the common
   leading key with INCLUDE MISSING. If however there are
   additional predicates for the OR clause, still skip combining
   the subarms, and try generating multiple indexes to be used
   in OR-index path.
 - fix another bug that if CBO-based advise is used, we lost that
   information when trying to combine multiple subarms of OR

Change-Id: I7bcb473d8cf79d1b4b7239b745a76cf2d76b6217
Reviewed-on: https://review.couchbase.org/c/query/+/192487
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>