MB-57647 Do not use covering index scan for system indexes
The fix for MB-55993 introduced a new secondary index for "bucket"
in system:dictionary (as well as a number of other system keyspaces).
Certain queries may end up using a covering index scan on this new
index, however covering index scans are not properly implemented for
system indexes, which results in incorrect result.

The fix is to not generate covering index scans for system indexes.

Change-Id: I2be35b4e08c5a235fe487b083a7a7bb88c8dc556
Reviewed-on: https://review.couchbase.org/c/query/+/193195
Reviewed-by: Donald Haggart <donald.haggart@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>