MB-57680 With CBO do not always pick index scan with order
When CBO is in effect, and there are multiple indexes that are
sargable, CBO considers intersect scan among the multiple indexes,
and during this process, if there is an index that provides order,
either through ORDER pushdown, or early order, currently CBO just
pick that index without cost consideration.

For this case there are two indexes, one provides early order, one
does not, however the index that does not provide order is much
more selective, due to predicates that are not sargable for the
other index.

The index that provides early order does not have EXACTSPANS flag
set (due to presence of other predicates), and thus the LIMIT
cannot be pushed to the index scan. In such cases it may not make
sense to always pick the index that provides the early order.

The fix is as follows:

 - modify the condition for early order for an index scan: if
   the index scan has EXACTSPANS, then all is good; otherwise we
   keep track of predicates that are not "covered" by the index
   scan, and if such extra predicates exist, do not perform early
   order (since in this case LIMIT cannot be pushed to the index
   scan, and there is no clear benefit for an early order)
 - when CBO considers intersect scan, if an index scan provides
   order (ORDER pushdown or early order), only pick that index
   unconditionally if the index scan also has EXACTSPANS;
   otherwise let cost be the criteria.
 - fix an issue with SORT cardinality estimation in case of early
   order, only use the LIMIT cardinality with EXACTSPANS.

Change-Id: Ic891523efd113083eedccb30fe8b1e5c252b55fa
Reviewed-on: https://review.couchbase.org/c/query/+/193247
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-on: https://review.couchbase.org/c/query/+/193251
Well-Formed: Restriction Checker