MB-57586 [BP 56816] Serialise pool refresh & keyspace addition
A pool refresh running concurrently with the addition of a new keyspace
to the cache could lead to caching of a bucket that has been closed by
the refresh logic. Such a bucket will not receive updates as its updater
will exit.

The solution here is to prevent the refresh from progressing whilst the
addition is in progress.

Change-Id: I6eba42e265ff98512f69042c4ed0f79589b435fa
Reviewed-on: https://review.couchbase.org/c/query/+/193343
Well-Formed: Restriction Checker
Reviewed-by: Marco Greco <marco.greco@couchbase.com>
Tested-by: Donald Haggart <donald.haggart@couchbase.com>