MB-57673 [BP 56816] Serialise pool refresh & keyspace addition
A pool refresh running concurrently with the addition of a new keyspace
to the cache could lead to caching of a bucket that has been closed by
the refresh logic. Such a bucket will not receive updates as its updater
will exit.

The solution here is to prevent the refresh from progressing whilst the
addition is in progress.

Change-Id: I6504396addb64b45b488723e3f1f96fbe287a576
Reviewed-on: https://review.couchbase.org/c/query/+/193380
Reviewed-by: Marco Greco <marco.greco@couchbase.com>
Well-Formed: Restriction Checker
Tested-by: Donald Haggart <donald.haggart@couchbase.com>