MB-57716 Add DEFAULT as keyword
For (native) recursive CTE implementation, there is an optional
CYCLE clause that includes the DEFAULT keyword:
   CYCLE <exprs> SET <expr> TO <expr> DEFAULT <expr>

A potential issue with adding DEFAULT as keyword is that
"default" is already used in other places without it being a
keyword, e.g. as the default namespace name. Thus adding
DEFAULT as keyword would normally require such uses to be
quoted in backticks.

To minimize impact to existing users we will add DEFAULT as a
special keyword that still allows the normal uses of "default"
as (equivalent to) an identifier.

Change-Id: I90549c539957f1c84077f0a633455ab586fc1333
Reviewed-on: https://review.couchbase.org/c/query/+/193456
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Donald Haggart <donald.haggart@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>