MB-57729 Do not perform string replacement on the entire array key
The way the index advisor generates an array key can be problematic
if the unnest alias is a suffix of the keyspace alias, due to the
process of string replacement of the generated array index key.
The purpose of the string replacement is to remove the keyspace
prefix for the array index binding expression. E.g.:
  ALL ARRAY `us`.`type` FOR us IN `s`.`a1` END
should get converted to:
  ALL ARRAY `us`.`type` FOR us IN `a1` END

However since the string replacement is done on the entire key
string it accidentally replaced the binding variable as well,
which is wrong.

The fix is to only perform string replacement on the binding
expression.

Change-Id: I3302ad2a82c562bff62d1ea41f84abe70fd8e281
Reviewed-on: https://review.couchbase.org/c/query/+/193502
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>