MB-56816 Refine fix
Allow keyspace loading to retry when a refresh has occurred whilst it
was busy, being careful to not retry if the refresh was driven by
itself.

Also improve updater logging with a identifier in addition to bucket
address for the cases where a retry may reuse a bucket, allowing clearer
interpretation of the messages relating to possibly two overlapping (one
closing, one starting) updaters for the bucket.

Change-Id: I5fb13345bf654c09e589627f4f05fb2b02e248e1
Reviewed-on: https://review.couchbase.org/c/query/+/193524
Reviewed-by: Marco Greco <marco.greco@couchbase.com>
Tested-by: Donald Haggart <donald.haggart@couchbase.com>