MB-57781 Proper save/restore this.node on builder
When extended attribute was implemented in Vulcan (5.5) a new
field (this.node) is added to builder structure such that we
can easily traverse all expressions of the current (sub)statement
and look for references to extended attributes. The issue is
that the field (this.node) is not always properly saved/restored
when we switch substatements in the planner.

The fix includes:

 - for algebra.Select, add save/restore of this.node, also setup
   this.node = stmt, similar to other DML statements (DELETE,
   MERGE, UPDATE)
 - for set operations (UNION/INTERSECT/EXCEPT) add save/restore
   of this.node when switching subarms of the set operations
 - for algebra.Subselect, add save/restore of this.node

With these changes this.node should point to the current
substatement that is being planned.

Change-Id: Ie3d353df143467ba61c782a4df7a84683afda03e
Reviewed-on: https://review.couchbase.org/c/query/+/193605
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-on: https://review.couchbase.org/c/query/+/193641
Well-Formed: Restriction Checker