Use same plan flags for ORDER
With the fix/backport of MB-57587 in 7.1.5 and 7.2.1, we now
use flags in the plan operator ORDER in 7.1.5 and 7.2.1 as
well. If we take a plan generated in 7.1.5/7.2.1 and interpret
it in 7.5/7.6, we need to make sure the flags have the same
meaning.

Change the order of the flags definition such that it matches
the flag value in earlier versions (7.1.5/7.2.1).

Change-Id: Ic1d2981e264275cfa89d4392234a496a9f8fd50c
Reviewed-on: https://review.couchbase.org/c/query/+/193621
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>