MB-57863 Proper formalization of query containing only an ExpressionTerm
For a query containing a single "source", which could be a KeyspaceTerm,
an ExpressionTerm, or a SubqueryTerm, the formalization process can
prefix the alias to any unprefixed expression, e.g.:

  SELECT month FROM calendar c WHERE ...

becomes

  SELECT c.monty FROM calendar c WHERE ...

this process requires proper setting of 'this.keyspace' on Formalizer.

The issue with this defect is that if we are using an ExpressionTerm
this.keyspace is not properly setup. It is somewhat surprising that the
issue has not surfaced much earlier since the issue existed since
ExpressionTerm was introduced back in version 5.0.

The fix is to properly set up this.keyspace for ExpressionTerm. Also
modified SubqueryTerm such that the logic is the same between
ExpressionTerm and SubqueryTerm.

Change-Id: I05c1fc1d92beb8b2c1c0d8ff2ead24fe84fffb32
Reviewed-on: https://review.couchbase.org/c/query/+/193869
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-on: https://review.couchbase.org/c/query/+/193871
Well-Formed: Restriction Checker