MB-57900 Proper check for max limit of number of spans when composing multiple spans
We have a limit of FULL_SPAN_FANOUT (8192) when generating multiple index spans,
in case of OR clause and/or IN clause. However in case of this bug, without the
fix for MB-51114, we could be generating a larger number of spans, i.e. the limit
is not properly followed.

The fix is to slightly modify the check for FULL_SPAN_FANOUT when combining
index spans from multiple (composite) index keys. Also need to keep track of
the index key that has too many spans, and use _WHOLE_SPANS if it is the first
index key.

Change-Id: Id120534d8b80615c581896d3a4dc3529e0043fe1
Reviewed-on: https://review.couchbase.org/c/query/+/194019
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>