MB-57901 Limit number of spans generated from a long IN-list
The planner has an upper limit of number of index spans generated
to avoid potential performance issues in the index scan. However
currently this limit is not enforced if an IN-list has too many
elements, resulting in number of spans exceeding the limit.

Add logic in handling of IN-list to check the number of elements
in the IN-list, and if it exceeds the limit, generate a single
index span with ARRAY_MIN/ARRAY_MAX as the bounds, and apply the
IN-list predicate post index scan.

Change-Id: I063c1afc9b6c4e66ab9c50d16c53ed7589a509d0
Reviewed-on: https://review.couchbase.org/c/query/+/186519
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>
Reviewed-on: https://review.couchbase.org/c/query/+/194021
Well-Formed: Restriction Checker
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>