MB-57961 Add multi-byte aware string functions.
Standard string functions operate in bytes which is a high-performance
approach but is really limited to single-byte character data for
practical utility.  This adds "mb_" prefixed versions of string functions
that accept or return string positions which operate on characters
(hence multi-byte aware).  Inevitably they will not perform at the same
level as the byte counterparts with the deficit proportional to the
magnitude of the positions/length.

Also cleaned-up and pared down comments in the string function source file.

Change-Id: I35af43d9cd57634e2d339c5bdb5f77af4d28309f
Reviewed-on: https://review.couchbase.org/c/query/+/194078
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Marco Greco <marco.greco@couchbase.com>