MB-57904 Proper handle of WITH correlation in encoded_plan
The earlier fix for this bug calls AddCorrelatedIdentifiers()
function for correlation references in WITH clause. This
causes issues if we use encoded_plan for transfer of query
plan (e.g. for prepared statements), since the function
currently returns an error if the reference is not found in
the "allowed" list. For handling of "CORRELATED " subquery
in the parser we don't have references of the parent query
thus triggering this error.

The fix is to check for the flag in the formalizer and
use a predefined flag in case it is not found in the
allowed list.

Change-Id: I860ee0663453da17825d58e8f120d39a90276ef3
Reviewed-on: https://review.couchbase.org/c/query/+/194105
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>