MB-57240 Do not ignore error from formalization
The previous fix introduced a new function, CheckSetCorrelated()
which uses a new formalizer to detect and mark correlation in
a CORRELATED subquery. Previously if an (unexpected) error
occurs from the formalizer, we just ignore the error and mark
everything as correlated. This may be problematic in some cases
since the formalizer is a "mapper", i.e. it modifies expressions,
and in case of error it can map an expression to nil. If an
error occurs and some expression is reset (mapped) to nil, it
may cause issues downstream.
Thus we should not ignore an error from formalizer, and error
out if it occurs rather than attempt to continue.

Change-Id: I0416ce8f03d0068462b202bab0e506ef4395fabe
Reviewed-on: https://review.couchbase.org/c/query/+/194108
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>