MB-57964 Add limit on completed_requests plan size
This adds a setting that sets the limit on the encoded plan size stored
in completed_requests.  This setting is limited to 20 MiB - 128 KiB (to
allow for the rest of the record) so entries don't exceed the document
size limits.  The default is 256 KiB.

Plans that exceed the limit are replaced with a message, e.g.

  "plan": {
    "WARNING": "Plan (2.497 KiB) exceeds maximum permitted (1.953 KiB) size."
  }

If the maximum size is set to 0, plans are replaced with:

  "plan": {
    "WARNING": "Plan inclusion disabled."
  }

In the UI the plan appears blank and the plan text shows the message.

If a limit < 0 or > the maximum is specified it is silently converted to
the maximum.

Change-Id: I4373d55bd45570e2a7d49edd29ed3faff26d5a99
Reviewed-on: https://review.couchbase.org/c/query/+/194204
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Marco Greco <marco.greco@couchbase.com>