MB-53384 Use result cache for covered index scan on primary index only when there is no spans
The fix for MB-53384 allows a primary index to be used inside a
correlated subquery. This is handled by a result cache in the
Fetch operator in case of non-covering index scans and IndexScan3
operator in case of covering idnex scan.

The issue is we could actually use the primary index in a covering
index scan with an index span, in which case we are not getting
the full primary scan result. In such cases we should not be using
the result cache.

The fix is to check for index spans and only allow VALUED, FULL or
WHOLE spans for using the result cache.

Change-Id: If980fba6bcd8ee468b6c782bf8a96bd9fe68d959
Reviewed-on: https://review.couchbase.org/c/query/+/195286
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>