MB-58157 Make sure cached result values have appropriate "parent" values
When a correlated subquery is being executed, the values for the
correlation references are passed during execution time as
"parent" value. If result cache is used in either IndexScan3 or
Fetch operator (after fix of MB-53384), we execute the underlying
operator once and cache the result, and use the cached result in
subsequent executions.

The issue is that currently the cached result contains the
"parent" value of the very first execution, and thus when the
cached result is used, we did not get the proper "parent" value
for subsequent executions.

The fix is to not cache the AnnotatedValue being returned from
the operator, rather cache the "immediate" value: in case of
IndexScan3, we cache the IndexEntries, and in case of Fetch,
we cache the value representing the fetched document. Then
we construct the AnnotatedValue to be returned from that, using
the current "parent" value in the process. This ensures that
we use the appropriate "parent" value for each execution.

Change-Id: I1ca44dc09fd098bfd176008149836ef156f16c9c
Reviewed-on: https://review.couchbase.org/c/query/+/195627
Reviewed-by: Donald Haggart <donald.haggart@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>